cleanup and ppv2 only supports pex v1
