+++ packages/core/src/modules/presentation-exchange/models/index.ts
@@ -1,1 +1,7 @@
-export * from './PresentationSubmission'
+export * from './PexCredentialsForRequest'
+import type { PresentationDefinitionV1, PresentationDefinitionV2, PresentationSubmission } from '@sphereon/pex-models'
+
+type PresentationDefinition = PresentationDefinitionV1 | PresentationDefinitionV2
+
+// Re-export some types from sphereon library
+export type { PresentationDefinitionV1, PresentationDefinitionV2, PresentationSubmission, PresentationDefinition }
