+++ packages/core/src/modules/proofs/formats/presentation-exchange/PresentationExchangeProofFormat.ts
@@ -1,21 +1,36 @@
 import type {
   InputDescriptorToCredentials,
-  PresentationDefinition,
   PexCredentialsForRequest,
+  PresentationDefinitionV1,
+  PresentationSubmission,
 } from '../../../presentation-exchange'
-import type {
-  PresentationExchangePresentation,
-  PresentationExchangeProposal,
-  PresentationExchangeRequest,
-} from '../../models/v2'
+import type { W3cJsonPresentation } from '../../../vc/models/presentation/W3cJsonPresentation'
 import type { ProofFormat } from '../ProofFormat'
+
+export type PresentationExchangeProposal = PresentationDefinitionV1
+
+export type PresentationExchangeRequest = {
+  options?: {
+    challenge?: string
+    domain?: string
+  }
+  presentation_definition: PresentationDefinitionV1
+}
+
+export type PresentationExchangePresentation =
+  | (W3cJsonPresentation & {
+      presentation_submission: PresentationSubmission
+    })
+  // NOTE: this is not spec compliant, as it doesn't describe how to submit
+  // JWT VPs but to support JWT VPs we also allow the value to be a string
+  | string
 
 export interface PresentationExchangeProofFormat extends ProofFormat {
   formatKey: 'presentationExchange'
 
   proofFormats: {
     createProposal: {
-      presentationDefinition: PresentationDefinition
+      presentationDefinition: PresentationDefinitionV1
     }
 
     acceptProposal: {
@@ -26,7 +41,7 @@ export interface PresentationExchangeProofFormat extends ProofFormat {
     }
 
     createRequest: {
-      presentationDefinition: PresentationDefinition
+      presentationDefinition: PresentationDefinitionV1
       options?: {
         challenge?: string
         domain?: string
