+++ packages/core/src/modules/proofs/models/index.ts
@@ -1,4 +1,3 @@
 export * from './ProofAutoAcceptType'
 export * from './ProofState'
 export * from './ProofFormatSpec'
-export * from './PresentationSubmission'
