--- packages/core/src/modules/proofs/models/v2/index.ts
@@ -1,1 +1,0 @@
-export * from './exchange'
