+++ packages/core/src/modules/proofs/protocol/v2/__tests__/fixtures.ts
@@ -1,3 +1,5 @@
+import type { InputDescriptorV1 } from '@sphereon/pex-models'
+
 export const TEST_INPUT_DESCRIPTORS_CITIZENSHIP = {
   constraints: {
     fields: [
@@ -7,4 +9,5 @@ export const TEST_INPUT_DESCRIPTORS_CITIZENSHIP = {
     ],
   },
   id: 'citizenship_input_1',
-}
+  schema: [{ uri: 'https://www.w3.org/2018/credentials/examples/v1' }],
+} satisfies InputDescriptorV1
